Renaming TransactionElement to NetworkElement
