+++ src/main/scala/org/bitcoins/core/protocol/NetworkElement.scala
@@ -1,13 +1,14 @@
-package org.bitcoins.core.protocol.transaction
+package org.bitcoins.core.protocol
 
 import org.bitcoins.core.util.BitcoinSUtil
 
+
 /**
- * Created by chris on 1/14/16.
- * Represents an element of a transction.
- * Examples would be inputs, outputs, scriptSigs, scriptPubKeys etc.
- */
-trait TransactionElement {
+  * Created by chris on 1/14/16.
+  * This represents a element that can be serialized to
+  * be sent over the network
+  */
+trait NetworkElement {
 
   /**
    * The size of the TransactionElement in bytes.
