+++ src/main/scala/org/bitcoins/core/protocol/script/ScriptPubKey.scala
@@ -3,7 +3,6 @@ package org.bitcoins.core.protocol.script
 import org.bitcoins.core.crypto.{ECFactory, ECPublicKey}
 import org.bitcoins.core.serializers.script.{RawScriptPubKeyParser, ScriptParser}
 import org.bitcoins.core.protocol._
-import org.bitcoins.core.protocol.transaction.TransactionElement
 import org.bitcoins.core.script.bitwise.{OP_EQUAL, OP_EQUALVERIFY}
 import org.bitcoins.core.script.constant._
 import org.bitcoins.core.script.crypto.{OP_CHECKMULTISIG, OP_CHECKMULTISIGVERIFY, OP_CHECKSIG, OP_HASH160}
@@ -13,7 +12,7 @@ import org.bitcoins.core.util.{BitcoinSLogger, BitcoinScriptUtil, Factory}
 /**
  * Created by chris on 12/26/15.
  */
-sealed trait ScriptPubKey extends TransactionElement with BitcoinSLogger {
+sealed trait ScriptPubKey extends NetworkElement with BitcoinSLogger {
 
   /**
    * Representation of a scriptSignature in a parsed assembly format
